TWINMIND-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pgvector==0.2.4

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Document Processing
pypdf==3.17.1
beautifulsoup4==4.12.2

# Image Processing
Pillow==10.1.0

# Task Queue & Cache
celery==5.3.4
redis==5.0.1

# LLM & ML
openai==1.3.0
langchain==0.1.0

# Audio Processing
pydub==0.25.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.0

# Type Hints
typing-extensions==4.8.0